# static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "1.0.0",
    "title": "CAPEL API-B2B",
    "description": "Microservicios para la gestión de datos relacionados con retailers y distribuidores",
    "contact": {
      "name": "Vladimir Reyes La O",
      "email": "vladimir.reyess@gmail.com"
    }
  },
  "paths": {
    "/api/v1/login": {
      "post": {
        "tags": [
          "Seguridad"
        ],
        "summary": "Login y autenticación de usuarios. Creación de token de accesos",
        "description": "Realiza la autenticación de usuarios, la creación de tokens de acceso y su gestión",
        "operationId": "login",
        "parameters": [
          {
            "in": "body",
            "name": "username",
            "example": "apicapel-username",
            "schema": {
              "type": "string"
            },
            "description": "Nombre de usuario registrado en mongoDB"
          },
          {
            "in": "body",
            "name": "password",
            "schema": {
              "type": "string"
            },
            "description": "Contraseña de usuario mongoDB"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "ACCESS FAILURE",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Acces Failure"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data": {
      "get": {
        "tags": [
          "Data"
        ],
        "summary": "Devuelve resultados enmarcados en un rango de fecha",
        "description": "Devuelve resultados enmarcados en un rango de fecha",
        "operationId": "data",
        "parameters": [
          {
            "in": "header",
            "name": "initialdate",
            "example": "%Y%m%d -- 20220601", 
            "schema": {
              "type": "string"
            },
            "description": "Fecha inicial periodo de interés"
          },
          {
            "in": "header",
            "name": "finaldate",
            "example": "%Y%m%d -- 20220601", 
            "schema": {
              "type": "string"
            },
            "description": "Fecha final periodo de interés"
          }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "summary": "Verifica que el objeto request esté bien conformado. Y devuelve mensaje inficativo del error existente",
            "description": "ERR_DATES_FORMAT | ERR_DATES_NOT_DEFINED | ERR_DIFF_DATE | ERR_FILTERS_EMPTY | ERR_RESPONSE_TOO_LARGE", 
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERR_DATES_FORMAT"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "summary": "Devuele errores del lado del servidor",
            "description": "ERR_INTERNAL", 
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERR_INTERNAL"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dailydata": {
      "get": {
        "tags": [
          "Data"
        ],
        "summary": "Devuelve la carga perteneciente a una data diaria(3 días)",
        "description": "Devuelve la carga perteneciente a una data diaria(3 días)",
        "operationId": "dailydata",
        "parameters": [
          {
            "in": "header",
            "name": "initialdate",
            "example": "%Y%m%d -- 20220601", 
            "schema": {
              "type": "string"
            },
            "description": "Fecha inicial periodo de interés"
          }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "summary": "Verifica que el objeto request esté bien conformado. Y devuelve mensaje inficativo del error existente",
            "description": "ERR_DATES_FORMAT | ERR_DATES_NOT_DEFINED | ERR_FILTERS_EMPTY | ERR_RESPONSE_TOO_LARGE", 
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERR_DATES_FORMAT"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "summary": "Devuele errores del lado del servidor",
            "description": "ERR_INTERNAL", 
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERR_INTERNAL"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/monthlydata": {
      "get": {
        "tags": [
          "Data"
        ],
        "summary": "Devuelve la carga perteneciente a una data mensual(30 días)",
        "description": "Devuelve la carga perteneciente a una data mensual(30 días)",
        "operationId": "monthlydata",
        "parameters": [
          {
            "in": "header",
            "name": "initialdate",
            "example": "%Y%m%d -- 20220601", 
            "schema": {
              "type": "string"
            },
            "description": "Fecha inicial periodo de interés"
          }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "summary": "Verifica que el objeto request esté bien conformado. Y devuelve mensaje inficativo del error existente",
            "description": "ERR_DATES_FORMAT | ERR_DATES_NOT_DEFINED | ERR_FILTERS_EMPTY | ERR_RESPONSE_TOO_LARGE", 
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERR_DATES_FORMAT"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "summary": "Devuele errores del lado del servidor",
            "description": "ERR_INTERNAL", 
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERR_INTERNAL"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }, 
    "/api/v1/filter": {
      "get": {
        "tags": [
          "Data"
        ],
        "summary": "Devuelve la carga perteneciente a un retailer en una fecha determinada, acotada por comuna, ciudad, marca",
        "description": "Devuelve la carga perteneciente a un retailer en una fecha determinada, acotada por comuna, ciudad, marca",
        "operationId": "filter",
        "parameters": [
          {
            "in": "header",
            "name": "initialdate",
            "example": "%Y%m%d -- 20220601", 
            "schema": {
              "type": "string"
            },
            "description": "Fecha inicial periodo de interés"
          },
          {
            "in": "header",
            "name": "finaldate",
            "example": "%Y%m%d -- 20220601", 
            "schema": {
              "type": "string"
            },
            "description": "Fecha final periodo de interés"
          },
          {
            "in": "header",
            "name": "retail",
            "example": "WALMART", 
            "schema": {
              "type": "string"
            },
            "description": "Retailer"
          },
          {
            "in": "header",
            "name": "comuna",
            "example": "SAN BERNARDO", 
            "schema": {
              "type": "string"
            },
            "description": "Comuna"
          },
          {
            "in": "header",
            "name": "marca",
            "example": "MADDERO", 
            "schema": {
              "type": "string"
            },
            "description": "Marca"
          },
          {
            "in": "header",
            "name": "ciudad",
            "example": "Nombre de ciudad", 
            "schema": {
              "type": "string"
            },
            "description": "Ciudad"
          }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "summary": "Verifica que el objeto request esté bien conformado. Y devuelve mensaje inficativo del error existente",
            "description": "ERR_DATES_FORMAT | ERR_DATES_NOT_DEFINED | ERR_FILTERS_EMPTY | ERR_RESPONSE_TOO_LARGE", 
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERR_DATES_FORMAT"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "summary": "Devuele errores del lado del servidor",
            "description": "ERR_INTERNAL", 
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERR_INTERNAL"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/filtercadena": {
      "get": {
        "tags": [
          "Data"
        ],
        "summary": "Devuelve la carga perteneciente a un retailer en una fecha determinada.",
        "description": "Devuelve la carga perteneciente a un retailer en una fecha determinada",
        "operationId": "filtercadena",
        "parameters": [
          {
            "in": "header",
            "name": "initialdate",
            "example": "%Y%m%d -- 20220601", 
            "schema": {
              "type": "string"
            },
            "description": "Fecha inicial periodo de interés"
          },
          {
            "in": "header",
            "name": "finaldate",
            "example": "%Y%m%d -- 20220601", 
            "schema": {
              "type": "string"
            },
            "description": "Fecha final periodo de interés"
          },
          {
            "in": "header",
            "name": "retail",
            "example": "WALMART", 
            "schema": {
              "type": "string"
            },
            "description": "Retailer"
          }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "summary": "Verifica que el objeto request esté bien conformado. Y devuelve mensaje inficativo del error existente",
            "description": "ERR_DATES_FORMAT | ERR_DATES_NOT_DEFINED | ERR_FILTERS_EMPTY | ERR_RESPONSE_TOO_LARGE", 
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERR_DATES_FORMAT"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "summary": "Devuele errores del lado del servidor",
            "description": "ERR_INTERNAL", 
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERR_INTERNAL"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/filtermarca": {
      "get": {
        "tags": [
          "Data"
        ],
        "summary": "Devuelve la carga perteneciente a una marca en una fecha determinada.",
        "description": "Devuelve la carga perteneciente a una marca en una fecha determinada",
        "operationId": "filtermarca",
        "parameters": [
          {
            "in": "header",
            "name": "initialdate",
            "example": "%Y%m%d -- 20220601", 
            "schema": {
              "type": "string"
            },
            "description": "Fecha inicial periodo de interés"
          },
          {
            "in": "header",
            "name": "finaldate",
            "example": "%Y%m%d -- 20220601", 
            "schema": {
              "type": "string"
            },
            "description": "Fecha final periodo de interés"
          },
          {
            "in": "header",
            "name": "marca",
            "example": "MADDERO", 
            "schema": {
              "type": "string"
            },
            "description": "Marca"
          }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "summary": "Verifica que el objeto request esté bien conformado. Y devuelve mensaje inficativo del error existente",
            "description": "ERR_DATES_FORMAT | ERR_DATES_NOT_DEFINED | ERR_FILTERS_EMPTY | ERR_RESPONSE_TOO_LARGE", 
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERR_DATES_FORMAT"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "summary": "Devuele errores del lado del servidor",
            "description": "ERR_INTERNAL", 
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERR_INTERNAL"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/filterciudad": {
      "get": {
        "tags": [
          "Data"
        ],
        "summary": "Devuelve la carga perteneciente a una ciudad en una fecha determinada.",
        "description": "Devuelve la carga perteneciente a una ciudad en una fecha determinada",
        "operationId": "filtermarca",
        "parameters": [
          {
            "in": "header",
            "name": "initialdate",
            "example": "%Y%m%d -- 20220601", 
            "schema": {
              "type": "string"
            },
            "description": "Fecha inicial periodo de interés"
          },
          {
            "in": "header",
            "name": "finaldate",
            "example": "%Y%m%d -- 20220601", 
            "schema": {
              "type": "string"
            },
            "description": "Fecha final periodo de interés"
          },
          {
            "in": "header",
            "name": "ciudad",
            "example": "Nombre de ciudad", 
            "schema": {
              "type": "string"
            },
            "description": "Ciudad"
          }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "summary": "Verifica que el objeto request esté bien conformado. Y devuelve mensaje inficativo del error existente",
            "description": "ERR_DATES_FORMAT | ERR_DATES_NOT_DEFINED | ERR_FILTERS_EMPTY | ERR_RESPONSE_TOO_LARGE", 
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERR_DATES_FORMAT"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "summary": "Devuele errores del lado del servidor",
            "description": "ERR_INTERNAL", 
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "ERR_INTERNAL"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}